+++ dsp/ptutil/include/Pwav.h
@@ -21,13 +21,13 @@
 #define PWAV_MAX_BUFFER_SIZE 10666
 
 /* pwav.cpp */
-int PT_DECLSPEC pwavNew(PT_HANDLE **, CSlout *, long, DWORD, int); 
-int PT_DECLSPEC pwavResetBufferSize(PT_HANDLE *, long, int);
-int PT_DECLSPEC pwavReadHeader(PT_HANDLE *, wchar_t *, wchar_t *, int, int, int *, long *, long *, int *);
-int PT_DECLSPEC pwavReadNextBuffer(PT_HANDLE *, int, long **, long *, long *, int, int);  
-int PT_DECLSPEC pwavGetBuffer(PT_HANDLE *, int, long **);
+int PT_DECLSPEC pwavNew(PT_HANDLE **, CSlout *, int32_t, DWORD, int); 
+int PT_DECLSPEC pwavResetBufferSize(PT_HANDLE *, int32_t, int);
+int PT_DECLSPEC pwavReadHeader(PT_HANDLE *, wchar_t *, wchar_t *, int, int, int *, int32_t *, int32_t *, int *);
+int PT_DECLSPEC pwavReadNextBuffer(PT_HANDLE *, int, int32_t **, int32_t *, int32_t *, int, int);  
+int PT_DECLSPEC pwavGetBuffer(PT_HANDLE *, int, int32_t **);
 int PT_DECLSPEC pwavGetTempFloatBuffer(PT_HANDLE *, realtype **);
-int PT_DECLSPEC pwavPlayBuffer(PT_HANDLE *, int, long); 
+int PT_DECLSPEC pwavPlayBuffer(PT_HANDLE *, int, int32_t); 
 int PT_DECLSPEC pwavGetDoneReading(PT_HANDLE *, int *);
 int PT_DECLSPEC pwavStop(PT_HANDLE *, int);
 int PT_DECLSPEC pwavCheckLegalFiles(PT_HANDLE *, wchar_t *, wchar_t *, int, int *);
@@ -45,12 +45,12 @@ int PT_DECLSPEC pwavSetupReadFromInput(PT_HANDLE *hp_pwav, int i_num_channels_in
 						   int i_num_channels_out, realtype r_samples_per_sec,
 						   int i_wave_in_dev, int i_wave_out_dev, int *ip_24_bit_flag);
 int PT_DECLSPEC pwavStopIO(PT_HANDLE *hp_pwav);
-int PT_DECLSPEC pwavReadNextIOBuffer(PT_HANDLE *hp_pwav, long **lpp_buffer_data);
-int PT_DECLSPEC pwavPlayIOBuffer(PT_HANDLE *hp_pwav, long *lp_buffer_data);
+int PT_DECLSPEC pwavReadNextIOBuffer(PT_HANDLE *hp_pwav, int32_t **lpp_buffer_data);
+int PT_DECLSPEC pwavPlayIOBuffer(PT_HANDLE *hp_pwav, int32_t *lp_buffer_data);
 int PT_DECLSPEC pwavQueueIOBuffer(PT_HANDLE *hp_pwav);
 int PT_DECLSPEC pwavStartIOProcessing(PT_HANDLE *hp_pwav);
 
 /* pwavUtil */
-int PT_DECLSPEC pwavCheckIfLegalSampFreq(long, int *);
+int PT_DECLSPEC pwavCheckIfLegalSampFreq(int32_t, int *);
 
 #endif //_PWAV_H
